changed CameraGPS to MediaGPS
