+++ .expo/settings.json
@@ -1,7 +1,7 @@
 {
   "hostType": "lan",
   "lanType": "ip",
-  "dev": true,
-  "minify": false,
-  "urlRandomness": "ed-g57"
+  "dev": false,
+  "minify": true,
+  "urlRandomness": "ik-x8u"
 }
