--- screens/CameraGPS.js
@@ -1,83 +1,0 @@
-import React, { Component } from 'react';
-import { Text, View, StyleSheet } from 'react-native';
-import { Constants, MapView, Location, Permissions } from 'expo';
-
-export default class App extends Component {
-  state = {
-    mapRegion: { latitude: 37.78825, longitude: -122.4324, latitudeDelta: 0.0922, longitudeDelta: 0.0421 },
-    locationResult: null,
-    location: {coords: { latitude: 37.78825, longitude: -122.4324}},
-    errorMessage: null,
-  };
-
-  componentDidMount() {
-    this._getLocationAsync();
-  }
-
-  _handleMapRegionChange = mapRegion => {
-    this.setState({ mapRegion });
-  };
-
-  _getLocationAsync = async () => {
-    const { status } = await Permissions.askAsync(Permissions.LOCATION);
-   if (status !== 'granted') {
-     this.setState({
-      errorMessage: 'Permission to access location was denied'
-     });
-     return;
-   }
-
-   const location = await Location.getCurrentPositionAsync({});
-   this.setState({ locationResult: JSON.stringify(location), location, });
- };
-
-  render() {
-
-    let text = 'Waiting...';
-
-    if (this.state.errorMessage) {
-      text = this.state.errorMessage;
-    } else if (this.state.location) {
-      text = this.state.locationResult;
-    }
-
-    return (
-      <View style={styles.container}>
-        <MapView
-          style={{ alignSelf: 'stretch', height: 400 }}
-          region={{ latitude: this.state.location.coords.latitude, longitude: this.state.location.coords.longitude, latitudeDelta: 0.0922, longitudeDelta: 0.0421 }}
-          onRegionChange={this._handleMapRegionChange}
-        >
-    <MapView.Marker
-    
-      coordinate={this.state.location.coords}
-      title="Confirm Task Location"
-      // description="Some description"
-    />
-        </MapView>
-      
-        <Text>
-          Location: {text}
-        </Text>
-      
-      </View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingTop: Constants.statusBarHeight,
-    backgroundColor: '#ecf0f1',
-  },
-  paragraph: {
-    margin: 24,
-    fontSize: 18,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    color: '#34495e',
-  },
-});
